Tweak scroll wheel handling.
Handlers can now do:
 wxTheApp->GetTopWindow()->ProcessEvent(event);
to ignore the underlying control's wheel behavior and send the event
back up to the EditorFrame, which will send it back out to the EditorCtrl.

See SearchEvtHandler for an example of the above.

EditorCtrl's event handler will now only grab the event if the mouse cursor
is within the bounds of the editor; otherwise it will allow other controls
that might be under the mouse cursor (such as the row of tabs) a chance
to respond to the wheel.

If nothing else in the frame responds, then the EditorCtrl gets the
event back.